+++ spectre-cli/src/main.rs
@@ -1,6 +1,5 @@
 use clap::{Parser, Subcommand, ValueEnum};
-pub use spectre::Error;
-use std::io::Cursor;
+use spectre::{Error, download::install_chrome};
 
 #[derive(Debug, Parser)]
 #[command(version, about)]
@@ -20,45 +19,19 @@ enum Browser {
     Chrome,
 }
 
-async fn download_chrome() -> Result<(), Error> {
-    println!("Downloading chrome...");
-    let url = "https://storage.googleapis.com/chrome-for-testing-public/135.0.7049.95/win64/chrome-win64.zip";
-    let response = reqwest::get(url).await?;
-
-    let bytes = response.bytes().await?;
-    let cursor = Cursor::new(bytes);
-    let mut archive = zip::ZipArchive::new(cursor)?;
-
-    println!("Unzipping...");
-
-    for i in 0..archive.len() {
-        let mut file = archive.by_index(i)?;
-        // FIXME
-        let outpath = file.enclosed_name().unwrap();
-
-        if file.is_dir() {
-            std::fs::create_dir_all(outpath)?;
-        } else {
-            if let Some(p) = outpath.parent() {
-                if !p.exists() {
-                    std::fs::create_dir_all(p)?;
-                }
-            }
-            let mut out_file = std::fs::File::create(&outpath)?;
-            std::io::copy(&mut file, &mut out_file)?;
-        }
-    }
-
-    Ok(())
-}
-
+// TODO improve messages and add a quiet flag
 #[tokio::main]
 async fn main() -> Result<(), Error> {
+    let home_dir = match home::home_dir() {
+        Some(dir) => dir,
+        None => return Err(Error::FailedToGetHomeDir),
+    };
+
     let args = Args::parse();
     match args.command {
         Command::Download { browser } => match browser {
             Browser::Chrome => {
-                download_chrome().await?;
+                install_chrome(&home_dir).await?;
             }
         },
     }