+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes will be tracked in this file. This isn't just a combination of
 commit messages/pull requests but rather a hand curated list of notable changes.
 
+## [Unreleased]
+
+### Changed
+
+- Download chrome to `$HOME/.spectre/browsers` instead of the the current working directory.
+
 ## 0.1.0 - 2025-4-28
 
 ### Added